refactor: 🚚 rename user-sidebar to sidebar
